+++ projects/common/src/lib/services/storage/storage.service.ts
@@ -14,6 +14,7 @@ import { CryptoHelper } from '@common';
 import {
   addIdentity,
   deleteIdentity,
+  editNick,
   switchIdentity,
 } from './related/identity';
 import { deletePermission } from './related/permission';
@@ -131,6 +132,13 @@ export class StorageService {
     await createNewVault.call(this, password);
   }
 
+  async editNick(
+    decryptedIdentityId: string,
+    decryptedNewNick: string,
+  ): Promise<void> {
+    await editNick.call(this, decryptedIdentityId, decryptedNewNick);
+  }
+
   async addIdentity(data: {
     nick: string;
     privkeyString: string;
@@ -172,7 +180,7 @@ export class StorageService {
     const vaultJson = JSON.stringify(
       this.getBrowserSyncHandler().browserSyncData,
       undefined,
-      4
+      4,
     );
     return vaultJson;
   }
@@ -181,14 +189,14 @@ export class StorageService {
     this.assureIsInitialized();
 
     const isValidData = this.#allegedBrowserSyncDataIsValid(
-      allegedBrowserSyncData
+      allegedBrowserSyncData,
     );
     if (!isValidData) {
       throw new Error('The imported data is not valid.');
     }
 
     await this.getBrowserSyncHandler().saveAndSetFullData(
-      allegedBrowserSyncData
+      allegedBrowserSyncData,
     );
   }
 
@@ -223,7 +231,7 @@ export class StorageService {
   assureIsInitialized(): void {
     if (!this.isInitialized) {
       throw new Error(
-        'StorageService is not initialized. Please call "initialize(...)" before doing anything else.'
+        'StorageService is not initialized. Please call "initialize(...)" before doing anything else.',
       );
     }
   }
@@ -238,13 +246,13 @@ export class StorageService {
     return CryptoHelper.encrypt(
       value,
       browserSessionData.iv,
-      browserSessionData.vaultPassword
+      browserSessionData.vaultPassword,
     );
   }
 
   async decrypt(
     value: string,
-    returnType: 'string' | 'number' | 'boolean'
+    returnType: 'string' | 'number' | 'boolean',
   ): Promise<any> {
     const browserSessionData =
       this.getBrowserSessionHandler().browserSessionData;
@@ -256,7 +264,7 @@ export class StorageService {
       value,
       returnType,
       browserSessionData.iv,
-      browserSessionData.vaultPassword
+      browserSessionData.vaultPassword,
     );
   }
 
@@ -264,7 +272,7 @@ export class StorageService {
     value: string,
     returnType: 'string' | 'number' | 'boolean',
     iv: string,
-    password: string
+    password: string,
   ): Promise<any> {
     const decryptedValue = await CryptoHelper.decrypt(value, iv, password);